Refactor: put classes in files, delete expand.php, add comments
* Put Parameter class and Template class in their own files
* Delete unused and ancient expand.php
* Add wayfinding comments and TODOs
* Continue adding braces and whitespace
* Replace with includeIfNew function with near-identical require_once
